installer: Fix bug with Desktop\config.xml
If you create a custom config file in the Desktop with the name
`config.xml` and execute the installer script, the `Get-ConfigFile`
function (introuced in https://github.com/mandiant/flare-vm/pull/531)
fails trying to copy the file to itself. Move the file instead of
copying it to avoid the duplication and that it fails.