[virtualbox] Add vbox-adapter-check.py
Add script to print the status of all internet adapters of all VMs in
VirtualBox. The script notifies using PyGObject if any dynamic analysis
VM (with `.dynamic` in the name) has an adapter whose type is not
allowed (internet access is undesirable for dynamic malware analysis).
Optionally change the type of the adapters with non-allowed type to
Host-Only.

The script has been tested in Debian 12 with GNOME 44.9.

Co-authored-by: Elliot Chernofsky <echernofsky@google.com>