[vbox] fix: Always try to remove the root snapshot
The current code excludes the root snapshot from deletion if a child is
protected. But it is possible to delete a root snapshot with a single
protected child.

It is not possible to delete the root snapshot if it has more than one
protected child. But this issue happens at every level, not only at the
root snapshot. For example (fail to delete 'Snapshot 1'):
```
ROOT
  ├─ Snapshot 1
  │   ├─ PROTECTED 1
  │   └─ PROTECTED 2
  └─ Snapshot 2

vbox-clean-snapshots.py FLARE-VM --protected "ROOT,PROTECTED" --root_snapshot ROOT
```

This inconsistent behavior is confusing and complicates the code
unnecessarily. Instead, try to always delete non protected snapshots
(including the root snapshot) and improve the rendered error output.

In the error output, include the command as a string (instead of a list)
as this allows to copy-paste the command to manually execute it.
Simplify also the long stderr from VBoxManage, by using only the first
`VBoxManage: error:` line to prevent using the long VBoxManage help
message or noisy information like the details and context.

Remove the use of `check` in `subprocess.run` (triggering an exception
on error) as it can be handle without exceptions.