[config] Remove libraries-extra.python3.vm
libraries-extra.python3.vm used to install the XRefer dependencies and
was removed with the introduction of the ida.plugin.xrefer.vm package in
https://github.com/mandiant/VM-Packages/pull/1207

We already have ida.plugin.xrefer.vm in the default configuration, so
libraries-extra.python3.vm is not needed. libraries-extra.python3.vm
will be removed from MyGet soon, not being available anymore.