Enhance FLARE-VM Installer
As part of the default config there are chocolatey packages that were not
displayed previously.
Only display packages sorted by category that are relevant to RE, excluded
hardcoded categories that belong to COMMANDO-VM.
Add a listbox that displays additional packages that are either chocolatey
packages or belong to excluded categories typically used in Commando-VM.
Add a new functionality to search across chocolatey packages including FLARE-VM
packages.

Add links for the choco community and VM-Packages wiki
Improve font to make it bigger.
Validate that the package didn't exist in the listBox before
adding a new one.
Disable the button Add Package if the text in the textbox changes
when it is enabled.
Allow to add additional packages to the listBox that are displayed categorized by categories.
Remove trailing whitespaces from the package textbox

Add new installer screenshot to the README

Change the description of the package selection in the README