Improve functionality to select and and new packages
Add links for the choco community and VM-Packages wiki
Improve font to make it bigger.
Validate that the package didn't exist in the listBox before
adding a new one.
Disable the button Add Package if the text in the textbox changes
when it is enabled.
Allow to add additional packages to the listBox that are displayed
as textboxes.