[vbox, config] Decouple FLARE-EDU
Separate the build process for the FLARE-EDU VM by removing the
specialized .EDU snapshot configuration from the `win10_flare-vm.yaml`
file and moving it to `virtualbox/configs/win10_flare-vm-edu.yaml`.

Add `microsoft-office.vm` to the default configuration, as it is an
useful analysis tool and simplifies `win10_flare-vm.yaml`. With this new
configuration, `win10_flare-vm.yaml` needs to used a custom config if
the installation of `microsoft-office.vm` is not desired.