Add a window that checks for preinstall checks
A new window will automatically check for preinstall checks, which
will allow the user to continue if the checks are not mandatory.
Two new global variables `$mandatoryChecksPassed` and `$checksPassed` are created
to determine if any of the mandatory or optional checks are passed or not.
If the mandatory checks are not passed the installation will display a message that
cannot continue..
New functions have been created for each of the preinstall checks.
Each function contains a try/catch to capture exceptions
Both in cli or gui mode the same functions must be invoked.
Install gui window is displayed inmediatelly before Boxstarter is installed
Preload gui controls of the custom configuration window
Add a checkbox in case some non-mandatory checks are not successful
Add a second checkbox to confirm a snapshot has been taken
Only display the first checkbox if some checks were not succesful