Add a window that checks for preinstall checks
A new window will automatically check for preinstall checks, which
will allow the user to continue if the checks are not mandatory.
Two new global variables `$mandatoryChecksPassed` and `$checksPassed` are created
to determine if any of the mandatory or optional checks are passed or not.
If the mandatory checks are not passed the installation will exit.
New functions have been created for each of the preinstall checks.
Both in cli or gui mode the same functions must be invoked.