[CI] Add workflow to release vbox executables
This commit introduces a new GitHub Actions workflow (`build-vbox.yaml`)
to automate the building and releasing of the VirtualBox helper scripts.

The workflow triggers automatically when a new tag prefixed with `vbox-`
is pushed, and can also be triggered manually for testing. It builds
standalone Linux executables for `vbox-adapter-check.py` and
`vbox-clean-snapshots.py` using PyInstaller. As `vbox-adapter-check.py`
uses GObject Introspection, the `vbox-adapter-check` executable is only
compatible with Linux desktop environments using the GTK toolkit, such
as GNOME.  The generated binaries are then bundled into a `vbox.zip`
archive and attached to the corresponding GitHub release. An explanatory
message is included in the release body to prevent FLARE-VM users from
confusing these script releases with main FLARE-VM project releases.